pdfai/requirements.txt
# Web framework and ASGI server
fastapi>=0.78.0
uvicorn[gunicorn]

# LangChain and utilities
langchain
langchain-community
openai

# PDF and image handling
pymupdf
pypdf
pdf2image
pillow

# OCR tools (DO NOT add `tesseract-ocr` here, it must be installed via apt)
pytesseract

# Vector search and FAISS
faiss-cpu
chromadb

# Tokenization and file handling
tiktoken
python-multipart

# Ollama API
ollama
